requirements.txt
git+https://github.com/Ricky294/nputils.git
numpy==1.24.1
pandas==1.5.2
python-binance==1.0.16
plotly==5.11.0
tqdm==4.64.1
PyYAML==6.0
TA-Lib==0.4.25
dash==2.7.1
requests==2.28.1
matplotlib
winotify
websocket-client
